requirements.txt

numpy==1.20.3
opencv-python-headless==4.5.3.56
Pillow==9.0.0
streamlit==1.3.1
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
matplotlib==3.5.1
scikit-learn
segmentation-models-pytorch